+++ javascript/node/selenium-webdriver/phantomjs.js
@@ -57,14 +57,6 @@ const CLI_ARGS_CAPABILITY = 'phantomjs.cli.args';
 
 
 /**
- * Default log file to use if one is not specified through CLI args.
- * @type {string}
- * @const
- */
-const DEFAULT_LOG_FILE = 'phantomjsdriver.log';
-
-
-/**
  * Custom command names supported by PhantomJS.
  * @enum {string}
  */
@@ -136,11 +128,16 @@ class Driver extends webdriver.WebDriver {
    *     capabilities.
    * @param {promise.ControlFlow=} opt_flow The control flow to use,
    *     or {@code null} to use the currently active flow.
+   * @param {string=} opt_logFile Path to the log file for the phantomjs
+   *     executable's output. For convenience, this may be set at runtime with
+   *     the `SELENIUM_PHANTOMJS_LOG` environment variable.
    */
-  constructor(opt_capabilities, opt_flow) {
+  constructor(opt_capabilities, opt_flow, opt_logFile) {
+    // TODO: add an Options class for consistency with the other driver types.
+
     var caps = opt_capabilities || capabilities.Capabilities.phantomjs();
     var exe = findExecutable(caps.get(BINARY_PATH_CAPABILITY));
-    var args = ['--webdriver-logfile=' + DEFAULT_LOG_FILE];
+    var args = [];
 
     var logPrefs = caps.get(capabilities.Capability.LOGGING_PREFS);
     if (logPrefs instanceof logging.Preferences) {
@@ -153,6 +150,11 @@ class Driver extends webdriver.WebDriver {
       if (level) {
         args.push('--webdriver-loglevel=' + level);
       }
+    }
+
+    opt_logFile = process.env['SELENIUM_PHANTOMJS_LOG'] || opt_logFile;
+    if (typeof opt_logFile === 'string') {
+      args.push('--webdriver-logfile=' + opt_logFile);
     }
 
     var proxy = caps.get(capabilities.Capability.PROXY);